readyset-mysql: Don't truncate UInt u64 to to i32
This reverts what's left of e9c66466342129d78fc7be5eca71c94cb75d04c2.

It is not clear what problem this hack was meant to solve, but reverting
it doesn't seem to break anything in tests (except an incorrect
implementation of mapping prepare resultset schema to `SqlType`,
resulting in benchmarks generating out of range values which are no
longer silently truncated)... and lets us support bigints larger than
2**32.

Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7789
Tested-by: Buildkite CI

+++ readyset-mysql/src/value.rs
@@ -10,7 +10,7 @@ pub(crate) fn mysql_value_to_dataflow_value(value: Value) -> ReadySetResult<DfVa
         ValueInner::Null => DfValue::None,
         ValueInner::Bytes(b) => DfValue::from(b),
         ValueInner::Int(i) => i.into(),
-        ValueInner::UInt(i) => (i as i32).into(),
+        ValueInner::UInt(i) => i.into(),
         ValueInner::Double(f) => DfValue::try_from(f)?,
         ValueInner::Datetime(_) => {
             if let Ok(ndt) = NaiveDateTime::try_from(value) {
